About to change a match statement.
About to change a match statement to an if statement. You can't compare
anything to a dynamic variable in a match statement (at least on
easily). So I'm going to change it to a good ole if statement. Once
done, the user should be able to customize their keys bindings.